refactor: replace KSelect usage with InstancioProvider.KSelect and remove unused RestValues object
