+++ src/main/kotlin/ru/virgil/spring/example/truck/TruckController.kt
@@ -4,7 +4,6 @@ import org.springframework.web.bind.annotation.*
 import ru.virgil.spring.example.box.BoxDto
 import ru.virgil.spring.example.box.BoxMapper
 import ru.virgil.spring.example.box.BoxService
-import ru.virgil.spring.example.system.rest.RestValues
 import ru.virgil.spring.tools.security.cors.GlobalCors
 import ru.virgil.spring.tools.util.Http.orNotFound
 import java.util.*
@@ -25,8 +24,8 @@ class TruckController(
 
     @GetMapping("/{truckUuid}/box")
     fun getBoxesByTruck(
-        @PathVariable truckUuid: UUID, @RequestParam(RestValues.PAGE) page: Int,
-        @RequestParam(RestValues.SIZE) size: Int,
+        @PathVariable truckUuid: UUID, @RequestParam page: Int,
+        @RequestParam size: Int,
     ): List<BoxDto> {
         val truck = truckService.get(truckUuid).orNotFound()
         val boxes = boxService.getAll(truck, page, size)
