+++ src/test/kotlin/ru/virgil/spring/example/test/BuyingOrderApiTest.kt
@@ -15,7 +15,6 @@ import ru.virgil.spring.example.order.BuyingOrderDto
 import ru.virgil.spring.example.order.BuyingOrderGenerator
 import ru.virgil.spring.example.order.BuyingOrderMapper
 import ru.virgil.spring.example.roles.user.WithMockedUser
-import ru.virgil.spring.example.system.rest.RestValues
 import ru.virgil.spring.example.truck.TruckDto
 import ru.virgil.spring.tools.SpringToolsConfig.Companion.BASE_PACKAGE
 import org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.testSecurityContext
@@ -39,7 +38,7 @@ class BuyingOrderApiTest @Autowired constructor(
     @Test
     fun getAll() {
         buyingOrderGenerator.generateAndSave(100)
-        val buyingOrderDtoList: List<BuyingOrderDto> = mockMvc.get("/buying_order?${RestValues.PAGE}=$page&${RestValues.SIZE}=$pageSize") {
+        val buyingOrderDtoList: List<BuyingOrderDto> = mockMvc.get("/buying_order?page=$page&size=$pageSize") {
             with(testSecurityContext())
         }.andExpect {
             status { isOk() }
@@ -62,8 +61,8 @@ class BuyingOrderApiTest @Autowired constructor(
     fun getTruckByOrder() {
         val buyingOrders = buyingOrderGenerator.generateAndSave(100)
         val uri = UrlBuilder.fromString("/buying_order/${buyingOrders.random().uuid}/truck")
-            .addParameter(RestValues.PAGE, page.toString())
-            .addParameter(RestValues.SIZE, pageSize.toString())
+            .addParameter("page", page.toString())
+            .addParameter("size", pageSize.toString())
             .toString()
         val truckDtoList: List<TruckDto> = mockMvc.get(uri) {
             with(testSecurityContext())
